// generated/src/app/api-models/report-definition.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ReportDefinition {
  keywordName: string;
  reportDesc: string;
  reportCategory: string;
  serverClientPrtInd: string;
  storedProcedureId: string;
  workfileTableName: string;
  parentJobIdInd: string;
  toolUsed: string;
  filterAllowInd: string;
  sortAllowInd: string;
  reprintAllowInd: string;
  userRestrictInd: string;
  parameter1: string;
  parameter2: string;
  parameter3: string;
  parameter4: string;
  parameter5: string;
  parameter6: string;
  parameter7: string;
  parameter8: string;
  parameter9: string;
  parameter10: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}